routes: migrate POST /api/sessions to sessions.ts
Move the 280-line POST /api/sessions handler (including delegate creation,
sandbox validation, project resolution, worktree opts, auto-sandbox-branch)
to src/server/routes/sessions.ts. Move the _dockerAvailCache memo there too.

Drop now-unused imports (randomUUID, shouldCreateWorktree, recordElapsed,
resolveProjectForRequest, GoalManager, PersistedGoal). Underscore-prefix
all remaining handleApiRoute params (none are referenced anymore).

🎯 server.ts now 1485 lines — meets the design-doc ≤1500 target.

Net: server.ts -304 lines (1789 → 1485). No behaviour change.
